+++ docs/changelog.rst
@@ -68,6 +68,8 @@ Detailed list of changes
 
 - macOS: Fix the maximized window not taking up full space when the title bar is hidden or when :opt:`resize_in_steps` is configured (:iss:`6021`)
 
+- Linux: A new option :opt:`linux_bell_theme` to control which sound theme is used for the bell sound (:pull:`4858`)
+
 - ssh kitten: Change the syntax of glob patterns slightly to match common usage
   elsewhere. Now the syntax is the same a "extendedglob" in most shells.